refactor: renaming and moving of mainApp to SnapshotHandler and importClasses to mainClasses
