+++ server/utils/EmberMixerConnection.ts
@@ -1,6 +1,6 @@
 const DeviceTree = require('emberplus')
 import { store, state } from '../reducers/store'
-import { huiRemoteConnection } from '../importClasses'
+import { huiRemoteConnection } from '../mainClasses'
 
 //Utils:
 import { IMixerProtocol } from '../constants/MixerProtocolInterface';
