+++ server/utils/SSLMixerConnection.ts
@@ -1,7 +1,7 @@
 //Node Modules:
 const net = require('net')
 import { store, state } from '../reducers/store'
-import { huiRemoteConnection } from '../importClasses'
+import { huiRemoteConnection } from '../mainClasses'
 
 //Utils:
 import { IMixerProtocol } from '../constants/MixerProtocolInterface'
